Show all episodes when more than 20 are available
